requirements.txt
asgiref==3.8.1
boto3==1.36.21
botocore==1.36.21
build==1.2.2.post1
Django==5.0.12
django-dotenv==1.4.2
django-mssql-backend==2.8.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
jmespath==1.0.1
mssql-django==1.5
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
pyodbc==5.2.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
s3transfer==0.11.2
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
